Update barley to sort on scales rather than datasets.
This was particularly necessary after vega/vega-dataflow@2488858.
Previously, with the variety and site datasets, the final Collector in
the pipeline would be empty, and thus the backing data assigned to
input.add. This caused the subsequent sort operator, called on the
Collector's backing data, to also operate over input.add and subsequently
used in the ordinal scales.